Refactored the code into smaller files
