+++ public/html/options.html
@@ -2,9 +2,10 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
+    <link rel="icon" type="image/svg+xml" href="img/infibox-logo.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Infibox Settings</title>
-    <link rel="stylesheet" href="options.css">
+    <link rel="stylesheet" href="css/options.css">
 </head>
 <body>
     <div class="container">
@@ -23,6 +24,6 @@
             <button type="submit">Save Settings</button>
         </form>
     </div>
-    <script src="options.js"></script>
+    <script src="js/options/index.js"></script>
 </body>
 </html>
