+++ public/html/popup.html
@@ -2,12 +2,12 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/infibox-logo.svg" />
+    <link rel="icon" type="image/svg+xml" href="img/infibox-logo.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Infibox - Popup</title>
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="popup.js"></script>
+    <script type="module" src="js/popup/index.js"></script>
   </body>
 </html>
